style: rename request_pool's mod
